requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
Brotli==1.0.9
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
dash==1.16.3
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
Django==3.1.1
django-plotly-dash==1.4.2
django-redis==4.12.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
dpd-components==0.1.0
easy-thumbnails==2.7
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
gunicorn==20.0.4
h5py==3.1.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras==2.4.3
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
plotly==4.11.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
service-identity==18.1.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
Twisted @ file:///D:/PG/myproject/dream/Twisted-20.3.0-cp38-cp38-win_amd64.whl
txaio==20.4.1
Werkzeug==1.0.1
whitenoise==5.2.0
xgboost==1.2.0
zope.interface==5.1.2
